# Interval: add default implementations for most methods