--- a/src/main/java/net/imglib2/Localizable.java
+++ b/src/main/java/net/imglib2/Localizable.java
@@ -48,11 +48,16 @@ public interface Localizable extends RealLocalizable
 {
 	/**
 	 * Write the current position into the passed array.
-	 * 
+	 *
 	 * @param position
 	 *            receives current position
 	 */
-	public void localize( int[] position );
+	default void localize( int[] position )
+	{
+		int n = numDimensions();
+		for ( int d = 0; d < n; d++ )
+			position[ d ] = getIntPosition( d );
+	}
 
 	/**
 	 * Write the current position into the passed array.
@@ -60,16 +65,24 @@ public interface Localizable extends RealLocalizable
 	 * @param position
 	 *            receives current position
 	 */
-	public void localize( long[] position );
+	default void localize( long[] position )
+	{
+		int n = numDimensions();
+		for ( int d = 0; d < n; d++ )
+			position[ d ] = getIntPosition( d );
+	}
 
 	/**
 	 * Return the current position in a given dimension.
-	 * 
+	 *
 	 * @param d
 	 *            dimension
 	 * @return dimension of current position
 	 */
-	public int getIntPosition( int d );
+	default int getIntPosition( int d )
+	{
+		return (int) getLongPosition( d );
+	}
 
 	/**
 	 * Return the current position in a given dimension.
@@ -79,4 +92,16 @@ public interface Localizable extends RealLocalizable
 	 * @return dimension of current position
 	 */
 	public long getLongPosition( int d );
+
+	@Override
+	default float getFloatPosition( int d )
+	{
+		return getLongPosition( d );
+	}
+
+	@Override
+	default double getDoublePosition( int d )
+	{
+		return getLongPosition( d );
+	}
 }