--- a/src/test/java/net/imglib2/RealLocalizableTest.java
+++ b/src/test/java/net/imglib2/RealLocalizableTest.java
@@ -11,13 +11,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -34,60 +34,59 @@
 
 package net.imglib2;
 
+import org.junit.Test;
+
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+
 /**
- * The {@link RealLocalizable} interface can localize itself in an n-dimensional
- * real space.
- * 
- * 
- * @author Stephan Preibisch
- * @author Stephan Saalfeld
+ * Tests {@link RealLocalizable}.
+ *
+ * @author Mathias Arzt
  */
-public interface RealLocalizable extends EuclideanSpace
+public class RealLocalizableTest
 {
-	/**
-	 * Write the current position into the passed array.
-	 * 
-	 * @param position
-	 *            receives current position
-	 */
-	default void localize( float[] position )
+	private final RealLocalizable localizable = new RealLocalizable()
 	{
-		int n = numDimensions();
-		for ( int d = 0; d < n; d++ )
-			position[ d ] = getFloatPosition( d );
+
+		@Override
+		public double getDoublePosition( int d )
+		{
+			return 4.2 + d;
+		}
+
+		@Override
+		public int numDimensions()
+		{
+			return 2;
+		}
+	};
+
+	@Test
+	public void testGetFloatPosition()
+	{
+		assertEquals( 5.2f, localizable.getFloatPosition( 1 ), 0 );
 	}
 
 	/**
-	 * Write the current position into the passed array.
-	 *
-	 * @param position
-	 *            receives current position
+	 * Tests {@link RealLocalizable#localize(double[])}.
 	 */
-	default void localize( double[] position )
+	@Test
+	public void testLocalizeWithDoubles()
 	{
-		int n = numDimensions();
-		for ( int d = 0; d < n; d++ )
-			position[ d ] = getDoublePosition( d );
+		double[] result = new double[ localizable.numDimensions() ];
+		localizable.localize( result );
+		assertArrayEquals( new double[] { 4.2, 5.2 }, result, 0 );
 	}
 
 	/**
-	 * Return the current position in a given dimension.
-	 * 
-	 * @param d
-	 *            dimension
-	 * @return dimension of current position
+	 * Tests {@link RealLocalizable#localize(float[])}.
 	 */
-	default float getFloatPosition( int d )
+	@Test
+	public void testLocalizeWithFloats()
 	{
-		return ( float ) getDoublePosition( d );
+		float[] result = new float[ localizable.numDimensions() ];
+		localizable.localize( result );
+		assertArrayEquals( new float[] { 4.2f, 5.2f }, result, 0 );
 	}
-
-	/**
-	 * Return the current position in a given dimension.
-	 * 
-	 * @param d
-	 *            dimension
-	 * @return dimension of current position
-	 */
-	public double getDoublePosition( int d );
 }