Most methods in the interface `Interval` can have a default implementation. 
These default implementations will prevent a lot of bugs and code duplication.
Implementing an Interval becomes as easy as overloading three methods:
```java
public class SimpleInterval implements Interval
{
	private final long[] min;
	private final long[] max;

	public SimpleInterval( long[] min, long[] max )
	{
		assert min.length == max.length;
		this.min = min;
		this.max = max;
	}

	@Override
	public long min( int d )
	{
		return min[ d ];
	}

	@Override
	public long max( int d )
	{
		return max[ d ];
	}

	@Override
	public int numDimensions()
	{
		return min.length;
	}
}
```