@maarzt 2019-01-11
I checked if adding this default implementations has any influence on performance.
Performance stays the same. 

The code I used for benchmarking: https://github.com/maarzt/imglib2/compare/master...benchmark-interval-default-implementations

```
################ Interval without changes ###############

Benchmark                                             Mode  Cnt        Score        Error  Units
DefaultIntervalBenchmark.abc                         thrpt    8   774081.514 ±  41143.894  ops/s
DefaultIntervalBenchmark.benchmarkMin                thrpt    8  2046413.208 ± 102335.106  ops/s
DefaultIntervalBenchmark.benchmarkRealMin            thrpt    8  2331398.003 ±  42055.958  ops/s
DefaultIntervalBenchmark.intervalContains            thrpt    8    52633.326 ±   2050.379  ops/s
DefaultIntervalBenchmark.intervalContainsRealPoints  thrpt    8  3888439.603 ± 461708.769  ops/s
DefaultIntervalBenchmark.iterateExtended             thrpt    8     9756.587 ±    274.826  ops/s
DefaultIntervalBenchmark.splitIntervall              thrpt    8     7148.196 ±   3052.784  ops/s

################ Interval with default implementations #############

Benchmark                                             Mode  Cnt        Score        Error  Units
DefaultIntervalBenchmark.abc                         thrpt    8   755103.282 ±  33699.039  ops/s
DefaultIntervalBenchmark.benchmarkMin                thrpt    8  2177057.950 ±  51290.428  ops/s
DefaultIntervalBenchmark.benchmarkRealMin            thrpt    8  2239097.764 ± 153554.875  ops/s
DefaultIntervalBenchmark.intervalContains            thrpt    8    53153.321 ±   1523.556  ops/s
DefaultIntervalBenchmark.intervalContainsRealPoints  thrpt    8  3931321.971 ± 212744.809  ops/s
DefaultIntervalBenchmark.iterateExtended             thrpt    8     9727.041 ±    391.607  ops/s
DefaultIntervalBenchmark.splitIntervall              thrpt    8     7222.925 ±   3215.695  ops/s
```