@ctrueden 2019-01-11
Awesome, I'm in big favor of migrating default implementations from abstract layers into interfaces wherever possible. Right now, implementing key ImgLib2 interfaces requires a lot of methods; and since ImgLib2 uses composition so much, you can't always extend a single abstract class to avoid it. Default methods for the win! :trophy: