@tpietzsch 2019-01-11
Yes, I also think it's time for a bit of cleaning up, using java 8 features.

However, the `RealInterval` methods should be implemented as default methods in `RealInterval` to avoid duplication