@ctrueden 2019-01-11
"Every default method should be pushed upstream as far as possible, but no farther" 😜