@tpietzsch 2019-01-30
@maarzt Could you move the `RealInterval` default methods to `RealInterval`?
(And run the benchmark again to see whether it makes a difference to have the more complicated  hierarchy with default methods inherited from different parents.)