@maarzt 2019-02-11
As requested I moved some default method implementations from `Interval` to `RealInterval` and `Dimensions`. I also added default method implementations to `Localizable` and `RealLocalizable`.
And I did the benchmark again.
```
**Before**
Benchmark                                             Mode  Cnt        Score        Error  Units
DefaultIntervalBenchmark.benchmarkMin                thrpt    8  2149053.201 ± 185376.861  ops/s
DefaultIntervalBenchmark.benchmarkRealMin            thrpt    8  2296187.559 ± 106860.403  ops/s
DefaultIntervalBenchmark.intervalContains            thrpt    8    50827.843 ±   1405.206  ops/s
DefaultIntervalBenchmark.intervalContainsRealPoints  thrpt    8  3585043.399 ± 695232.202  ops/s
DefaultIntervalBenchmark.iterateExtended             thrpt    8     3269.245 ±    136.902  ops/s
DefaultIntervalBenchmark.splitIntervall              thrpt    8     6929.482 ±   2940.917  ops/s

**After (with default method implementation**
Benchmark                                             Mode  Cnt        Score        Error  Units
DefaultIntervalBenchmark.benchmarkMin                thrpt    8  2150682.270 ± 141313.709  ops/s
DefaultIntervalBenchmark.benchmarkRealMin            thrpt    8  2314714.556 ±  82895.721  ops/s
DefaultIntervalBenchmark.intervalContains            thrpt    8    50241.894 ±   5093.540  ops/s
DefaultIntervalBenchmark.intervalContainsRealPoints  thrpt    8  3811206.108 ± 346780.226  ops/s
DefaultIntervalBenchmark.iterateExtended             thrpt    8     9459.974 ±    673.724  ops/s
DefaultIntervalBenchmark.splitIntervall              thrpt    8     6992.915 ±   2890.582  ops/s
```

Performance stays the same.
@tpietzsch I think it can be merged?